refactor: 질문 목록 데이터 검증 부분 함수로 분리
